+++ src/utils/create-element.utils.spec.ts
@@ -1,4 +1,4 @@
-import {createEmptyElement, createUneditableDiv} from './create-element.utils';
+import {createEmptyElement} from './create-element.utils';
 
 describe('create element', () => {
   it('should create empty element', () => {
@@ -6,12 +6,5 @@ describe('create element', () => {
 
     expect(empty.nodeName.toLowerCase()).toEqual('div');
     expect(empty.innerHTML).toEqual('\u200B');
-  });
-
-  it('should create a not editable element', () => {
-    const empty = createUneditableDiv();
-
-    expect(empty.nodeName.toLowerCase()).toEqual('div');
-    expect(empty.getAttribute('contenteditable')).toEqual('false');
   });
 });