+++ src/utils/create-element.utils.ts
@@ -8,17 +8,3 @@ export const createEmptyElement = ({
 
   return element;
 };
-
-// We do not want TextNode as direct child of the container.
-// If user types next to some HTML elements, such as hr and img, the resulting text is a TextNode with the container as parent.
-// <container>
-//   <img>
-// -> after use type
-// <container>
-//   text
-//   <img>
-export const createUneditableDiv = (): HTMLDivElement => {
-  const div: HTMLDivElement = document.createElement('div');
-  div.setAttribute('contenteditable', 'false');
-  return div;
-};