+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
 - preserve styles when removing links (except Firefox)
 - prevent text nodes - i.e. leaves - as direct children of the content editable when user is typing
+- `<img/>` and `<hr/>` do not need to be wrapped in a parent `contenteditable=false` anymore
 
 ### Fix
 
 - add link on Firefox
+- clone image on drag
 
 # 0.0.4-1 (2022-02-08)